+++ packages/common/src/index.js
@@ -1,3 +1,4 @@
 export * from "./jsonl.js";
+export * from "./logging.js";
 export * from "./protocol.js";
 export * from "./validation.js";